refactor(admin-ui): Phase2-P1-2 A5 — BookUploadsSection + PdfUploadTab を component 抽出 (機能変更なし) (GID 1214120401966111)
